separate post process parameters into own cpp and h files
